Add colored message for git errors
The motivation for this was using `git co foo` instead of `git co -b foo`. This obviously results in a pathspec error, but when going quickly, it's not obvious and changes can be made to the wrong branch.

The  error code of `1` covers this, as well as failed pushed which is sufficient for my use case.