Distinguish work tree and index changes
The untracked flag (`?`) is converted to an `A` for the work tree since it being untracked is implied. The untracked flag being present for the index is an artifact of the parsed format. Since a file in the index cannot be untracked, that flag is ignored.

I'm unclear on when the `!` flag occurs, but it may require a similar special case since it works like `?` in the parsed format.

Fixes #5